+++ it.sh
@@ -17,8 +17,8 @@ function gungnir-docker {
 
 function deploy {
     echo "Deploying Cluster"
-    git clone https://github.com/Comcast/codex.git 2> /dev/null || true
-    pushd codex/deploy/docker-compose
+    git clone https://github.com/xmidt-org/codex-deploy.git 2> /dev/null || true
+    pushd codex-deploy/deploy/docker-compose
     GUNGNIR_VERSION=$GUNGNIR_VERSION docker-compose up -d db db-init gungnir
     check $?
     popd
@@ -28,7 +28,7 @@ function deploy {
 gungnir-docker
 echo "Gungnir V:$GUNGNIR_VERSION"
 deploy
-go get -d github.com/Comcast/codex/tests/...
+go get -d github.com/xmidt-org/codex-deploy/tests/...
 printf "Starting Tests \n\n\n"
-go run github.com/Comcast/codex/tests/runners/travis -feature=codex/tests/features/gungnir/travis
+go run github.com/xmidt-org/codex-deploy/tests/runners/travis -feature=codex-deploy/tests/features/gungnir/travis
 check $?