+++ release_notes.md
@@ -1,1 +1,3 @@
-
+- Updated imports and versions for codex packages and webpa-common
+- Updated bascule 
+- Updated links to update new location for gungnir